# x/ibc/core/23-commitment/types: fix MerkleProof.Empty comparisons