--- a/x/ibc/core/23-commitment/types/bench_test.go
+++ b/x/ibc/core/23-commitment/types/bench_test.go
@@ -1,0 +1,15 @@
+package types
+
+import (
+	"testing"
+)
+
+func BenchmarkMerkleProofEmpty(b *testing.B) {
+	b.ReportAllocs()
+	var mk MerkleProof
+	for i := 0; i < b.N; i++ {
+		if !mk.Empty() {
+			b.Fatal("supposed to be empty")
+		}
+	}
+}