--- a/x/ibc/core/23-commitment/types/merkle.go
+++ b/x/ibc/core/23-commitment/types/merkle.go
@@ -270,8 +270,11 @@ func verifyChainedMembershipProof(root []byte, specs []*ics23.ProofSpec, proofs 
 }
 
 // Empty returns true if the root is empty
-func (proof MerkleProof) Empty() bool {
-	return proto.Equal(&proof, nil) || proto.Equal(&proof, &MerkleProof{}) || proto.Equal(&proof, &tmcrypto.ProofOps{})
+var blankMerkleProof = &MerkleProof{}
+var blankProofOps = &tmcrypto.ProofOps{}
+
+func (proof *MerkleProof) Empty() bool {
+	return proof == nil || proto.Equal(proof, blankMerkleProof) || proto.Equal(proof, blankProofOps)
 }
 
 // ValidateBasic checks if the proof is empty.