@odeke-em 2020-12-06
@marbar3778 I kindly wanted to ask: was this code that came from PR https://github.com/cosmos/cosmos-sdk/pull/6892 already in Tendermint? If so, please direct me to where, and we can then also make a fix in Tendermint. Thank you.