@odeke-em 2020-12-07
>This comparison is strictly in the SDK

Got it, thank you @marbar3778!